+++ NEWS.md
@@ -1,4 +1,6 @@
-# Version 1.0.9000.9114
+# Version 1.0.9000.9116
+* Added method `size` for objects of class `Samples`.
+* Added new custom `checkmate` function `check_length`.
 * Added `unique` flag to `assert_probabilities` checkmate custom functions.
 * Created a new vignette which describes how to use certain functions and features
   of `crmPack` after the major refactoring.